smart_campus_connect/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.4
PyMySQL==1.0.3
python-dotenv==1.0.0
